Moved contact details into its own seperate directory
