+++ src/components/ContactDetails/ContactDetails.jsx
@@ -1,6 +1,6 @@
 import React,{useState,useEffect} from 'react';
 import axios from 'axios';
-import './Contact.scss';
+import './ContactDetails.scss';
 const backendUrl = process.env.REACT_APP_BACKEND_URL;
 
 function ContactById(props) {
